pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lib"
version = "0.1.1"
description = "A collection of utility libraries for database, HTTP, Kafka, logging, config and utils"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "loguru>=0.7.0",
    "sqlalchemy>=1.4.0",
    "pandas>=1.3.0",
    "requests>=2.25.0",
    "confluent-kafka>=2.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "urllib3>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9",
]